refactor(contracts/ERC20Guild): Refactor proposals logic to allow multiple actions
The proposals struct and functions were changed to allow having proposals with multiple actions, the
action with votes over the requested for execution and higher votes than other proposals would be
executed when the proposal ends, which means that a proposal will have a winning action. Also the
voter now has to set the action for his vote and this cant be changed, the voter can only increase
the voting power in the proposal if it has voting power left to vote.