feat(contracts): add AnonREP721 token and AnonERC20Guild implementation
Add an implementation of an ERC721 token with vote commitments, the token keep track of the relation
between vote commitments and tokens, the token is snapshotted and can be mint and burned only by the
owner, very similar to the ERC20 snapshot rep token. This token is used by the AnonERC20Guild, this
guild allows the token holders to vote wth their tokens using ZK proof, by sharing a prrof of chain
of their vote using their tokens, this is possible thanks to the public voting commitment tracked on
the token.