refactor(contracts/dao/scheme): if all calls executed are empty the proposal is marked as rejected
Refactor the execution of proposals in Scheme, now a proposal is marked as rejected if the calls
batch executed are just empty calls (calls to address 0 and value 0). The WalletScheme was changed
requiring the second half of the proposals submitted to be empty, this means that for option 2 it
will mark the proposal as rejected, since the second half of the calls are just empty calls.