refactor(contracts/dao): refactor stake/redeem procedure in VotingMachine
Removed the daoBountyConst and now it only uses the thresholdConst to exponientially increase the
downstakes in new proposals,m tsrating from minimumDaoBounty, the more active boosted proposals the
more it will be requried to stake to get more boosted proposals. The Redeem process was simplified
as well, now we have only one redeem function, the bounty on the voting machine is always a fixed
dao bounty amount, also code that was not needed neither used was removed.